ABW-1949 fee breakdown (#390)
* ABW-1949.- Fee breakdown implemented.

* ABW-1949 - Lock fee and tip fixes.

* ABW-1949 - More improvements to Fee breakdown

* ABW-1949 - Logic for transactionFeeToLock changed. Unit tests for fee locks added.

* ABW-1949 - Strings replaced with copies from Crowdin.

* ABW-1949 - More copies added.

* ABW-1949 - Some fixes for padding corners and validation.

+++ app/src/main/java/com/babylon/wallet/android/domain/model/AccountWithResources.kt
@@ -28,9 +28,9 @@ data class Resources(
     val isNotEmpty: Boolean
         get() = fungibleResources.isNotEmpty() || nonFungibleResources.isNotEmpty()
 
-    fun hasXrd(minimumBalance: Long = 1L): Boolean = xrd?.let {
+    fun hasXrd(minimumBalance: BigDecimal = BigDecimal(1)): Boolean = xrd?.let {
         it.amount?.let { amount ->
-            amount >= BigDecimal(minimumBalance)
+            amount >= minimumBalance
         }
     } == true
 
@@ -69,7 +69,7 @@ data class ValidatorWithStakeResources(
     }
 }
 
-fun List<AccountWithResources>.findAccountWithEnoughXRDBalance(minimumBalance: Long) = find {
+fun List<AccountWithResources>.findAccountWithEnoughXRDBalance(minimumBalance: BigDecimal) = find {
     it.resources?.hasXrd(minimumBalance) ?: false
 }
